Moved genRandBytes to Utils
